third_party/WebKit/LayoutTests/fast/css/first-letter-before-hit-test.html

<!doctype html>
<script src="../../resources/js-test.js"></script>
<style>
#tab:first-letter {
  text-transform: uppercase;
}
#tab:before {
  content: "foo";
}
</style>
<div id="tab">content</div>

<script>
offset = document.getElementById('tab').getBoundingClientRect();
shouldBe('document.elementFromPoint(offset.left + 5, offset.top + 5).id', "'tab'");
</script>
