third_party/blink/web_tests/platform/generic/fast/css/first-letter-before-hit-test-expected.txt

PASS document.elementFromPoint(offset.left + 5, offset.top + 5).id is 'tab'
PASS successfullyParsed is true

TEST COMPLETE

content
